chore(output): improve code and refactor tests
