+++ prowler/lib/outputs/finding.py
@@ -82,7 +82,8 @@ class Finding(BaseModel):
     notes: str
     prowler_version: str = prowler_version
 
-    def __init__(self, provider, finding, output_options) -> "Finding":
+    @classmethod
+    def generate_output(self, provider, finding, output_options) -> "Finding":
         """generates the output for a finding based on the provider and output options
 
         Args:
@@ -107,8 +108,9 @@ class Finding(BaseModel):
             finding, provider.type, output_options
         )
         finding_output = self.generate_finding_output(provider, finding, output_data)
-        super().__init__(**finding_output)
+        return finding_output
 
+    @classmethod
     def generate_finding_output(self, provider, finding, output_data) -> dict:
         """generates the provider specific output for a finding
 
@@ -190,9 +192,10 @@ class Finding(BaseModel):
                 f"prowler-{provider.type}-{finding.check_metadata.CheckID}-{output_data['account_uid']}-{output_data['region']}-{output_data['resource_name']}"
             )
 
-            return output_data
+            finding_output = self(**output_data)
         except Exception as error:
             logger.error(
                 f"{error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
             )
-            return output_data
+        else:
+            return finding_output
